backprop into the transformer...
Add parameters to decide whether or not finetune the transformer
By default, train Bert with a low LR and very low WD to avoid it overreacting to the training data

Check bert_finetune when loading models.  Set it to false if not in the model so that we can preserve old models
Also, check bert_finetune for whether or not to use the foundation_cache when building a new model

Also backprop phobert.
TODO: xlnet and bart are also implemented, but need to be tested

Skip the first N steps for finetuning the transformer based on a parameter
Add an endpoint for the bert finetuning as well (another option would be separate initial phase and second phase values)
Add a separate argument for stage1_bert_finetuning
Includes functionality to finetune a limited number of layers

Instead of using the FoundationCache, we need to Load a new copy of
the bert_model if the bert_model was saved in the constituency file
(otherwise, the cached model would be clobered for any other users of
the model)

Add a single epoch test of the bert finetuning

Add a multiepoch test which checks that the weights are or aren't close to the final model in a manner expected by the bert training

TODO: if the endpoint is before the stage1 would be finished, presumably we could turn that into stage1_ vs non